+++ storage/indexer.go
@@ -8,6 +8,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"net"
+	"net/http"
 	"net/url"
 	"strings"
 	"sync"
@@ -100,7 +102,17 @@ func (i *Indexer) setupResolver() error {
 		return err
 	}
 
+	httpClient := http.Client{
+		Transport: &http.Transport{
+			DialContext: (&net.Dialer{
+				// Connect timeout.
+				Timeout: 20 * time.Second,
+			}).DialContext,
+		},
+	}
+
 	i.resolver = storageResolver{
+		client:               &httpClient,
 		artifactsPackagesURL: *baseURL.ResolveReference(&url.URL{Path: artifactsPackagesStoragePath + "/"}),
 		artifactsStaticURL:   *baseURL.ResolveReference(&url.URL{Path: artifactsStaticStoragePath + "/"}),
 	}