+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Reuse HTTP client when proxifying resolver requests. [#1147](https://github.com/elastic/package-registry/pull/1147)
+
 ### Deprecated
 
 ### Known Issues